sva-73: move query type function to types.js
