+++ src/queries/types.js
@@ -1,3 +1,5 @@
+import * as _ from 'lodash';
+
 export const QUERY_TYPES = {
   APP_USER_CONTENT: 'appUserContent',
   CATEGORIES: 'categories',
@@ -12,4 +14,18 @@ export const QUERY_TYPES = {
   PUBLIC_JSON_FILE: 'publicJsonFile',
   TOUR: 'tour',
   TOURS: 'tours'
+};
+
+/**
+ * Parse a query type from json string if one matches.
+ * @argument {string} input
+ * @return {string | undefined}
+ */
+export const getQueryType = (input) => {
+  const camelCaseInput = _.camelCase(input);
+  const availableTypes = [QUERY_TYPES.TOUR,
+    QUERY_TYPES.POINTS_OF_INTEREST,
+    QUERY_TYPES.NEWS_ITEM,
+    QUERY_TYPES.EVENT_RECORD];
+  return availableTypes.find((type) => type === camelCaseInput);
 };
